Move StewItemMixin functionality to ItemMixin, make StewItemMixin override finishUsing
